backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pikepdf==9.0.0
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.4.0
reportlab==4.2.5
PyMuPDF==1.26.5
pytesseract==0.3.13
python-magic==0.4.27